backend/requirements.txt
aiohttp==3.7.4.post0
aiohttp-cors==0.7.0
aiosqlite==0.17.0
alembic==1.6.5
appnope==0.1.2
async-timeout==3.0.1
asyncpg==0.23.0
attrs==21.2.0
backcall==0.2.0
chardet==4.0.0
decorator==5.0.9
Faker==8.10.3
flake8==3.9.2
greenlet==1.1.0
idna==3.2
iniconfig==1.1.1
ipython==7.25.0
ipython-genutils==0.2.0
jedi==0.18.0
Mako==1.1.4
MarkupSafe==2.0.1
matplotlib-inline==0.1.2
mccabe==0.6.1
mtpylon==0.0.1
multidict==5.1.0
mypy==0.910
mypy-extensions==0.4.3
packaging==21.0
parso==0.8.2
pexpect==4.8.0
pickleshare==0.7.5
pluggy==0.13.1
prompt-toolkit==3.0.19
ptyprocess==0.7.0
py==1.10.0
pyaes==1.6.1
pyasn1==0.4.8
pycodestyle==2.7.0
pyflakes==2.3.1
Pygments==2.9.0
pyparsing==2.4.7
pytest==6.2.4
pytest-asyncio==0.15.1
python-dateutil==2.8.2
python-dotenv==0.18.0
python-editor==1.0.4
rsa==4.7.2
six==1.16.0
SQLAlchemy==1.4.21
sqlalchemy2-stubs==0.0.2a5
text-unidecode==1.3
TgCrypto==1.2.2
toml==0.10.2
traitlets==5.0.5
typing-extensions==3.10.0.0
wcwidth==0.2.5
yarl==1.6.3
